# app/Docs/ProductRequestSchema.php
<?php

namespace App\Docs;

/**
 * @OA\Schema(
 *     schema="ProductRequest",
 *     @OA\Property(property="name", type="string"),
 *     @OA\Property(property="description", type="string"),
 *     @OA\Property(property="price", type="number")
 * )
 */
class ProductRequestSchema
{
}